// src/app/core/storage/indexed-db/indexed-db-storage-key.enum.ts
export enum TrudiIndexedDBStorageKey {
  MESSAGE = 'message',
  MESSAGE_MAIL_FOLDER = 'messageMailFolder',
  TASK = 'task',
  CONVERSATION = 'conversation',
  TASK_LIST = 'taskList',
  TASK_GROUP = 'taskGroup',
  TASK_FOLDER = 'taskFolder',
  TASK_DETAIL = 'taskDetail',
  TASK_CONVERSATIONS = 'taskConversations',
  TASK_ITEMS = 'taskItems',
  TASK_USER_CONVERSATION = 'taskUserConversation',
  CONVERSATIONS = 'conversations',
  CALENDAR_EVENTS = 'calendarEvents',
  CALENDAR_DASHBOARD = 'calendarDashboard',
  TENANTS_OWNERS_PT = 'tenantsOwnersPT',
  TENANTS_OWNERS_RM = 'tenantsOwnersRM',
  SUPPLIER = 'supplier',
  OTHER_CONTACT = 'otherContact',
  TENANT_PROSPECT = 'tenantProspect',
  OWNER_PROSPECT = 'ownerProspect',
  MAIL_FOLDER = 'mailFolder',
  APP_MESSAGE = 'appMessage',
  VOICE_MAIL = 'voiceMail',
  FACEBOOK = 'facebook',
  SMS = 'sms',
  WHATSAPP = 'whatsapp'
}

export enum TrudiIndexedDBIndexKey {
  TASK_ID = 'taskId',
  EVENT_ID = 'eventId',
  TASK_FOLDER_ID = 'taskFolderId',
  COMPANY_ID = 'companyId',
  DATE = 'date',
  TYPE = 'type',
  ID = 'id',
  PROPERTY_ID = 'propertyId',
  EXTERNAL_ID = 'externalId',
  USER_ID = 'userId',
  MAIL_BOX_ID = 'mailBoxId'
}